Analyze and fix multiple flaws in the project
This commit addresses several issues found during a comprehensive analysis of the project. The main goal was to fix critical bugs, improve the architecture, and update the documentation.

The following changes were made:

1.  **Updated Documentation:**
    -   The `README.md` was completely rewritten to reflect the current state of the project, which is a web application with a React frontend and a FastAPI backend.
    -   Added correct installation and execution instructions.

2.  **Fixed Critical Bug in Alert Configuration:**
    -   Implemented the `handleConfigChange` function in the frontend to correctly save alert configuration changes.
    -   The frontend now sends the updated configuration to the backend's `/api/alert_configs` endpoint.

3.  **Refactored Frontend Architecture:**
    -   The monolithic `index.tsx` file was broken down into smaller, reusable components (`App`, `CryptoCard`, `SettingsModal`, `AlertsPanel`).
    -   The frontend source code was reorganized into a standard `src` directory structure with a `components` sub-directory.
    -   The Vite configuration was updated to work with the new structure.

4.  **Addressed Minor Flaws:**
    -   Removed the unused `@google/genai` dependency from `package.json`.
    -   Added the `@vitejs/plugin-react` dependency, which is required for Vite to work with React.
    -   Replaced the hardcoded Bitcoin dominance value with a live value fetched from the CoinGecko API.
